+++ libs/components/forms/src/lib/modules/required-state/fixtures/template-driven.component.fixture.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+
+import { TestControlComponent } from './test-control.component.fixture';
+
+@Component({
+  imports: [FormsModule, TestControlComponent],
+  standalone: true,
+  template: `<sky-test-control [required]="required" [(ngModel)]="foobar" />`,
+})
+export class TemplateDrivenTestComponent {
+  public required = false;
+
+  protected foobar = '';
+}