+++ libs/components/layout/src/lib/modules/box/fixtures/box.component.fixture.ts
@@ -8,4 +8,5 @@ export class BoxTestComponent {
   public ariaLabel: string | undefined;
   public ariaLabelledBy: string | undefined;
   public ariaRole: string | undefined;
+  public showHeader = true;
 }